requirements.txt
python-dotenv
faiss-cpu
langchain
langchain-openai
langchain-chroma
langchain-text-splitters
langchain-anthropic
langchain-google-genai
langchain-google-firestore
firestore
tiktoken
sentence-transformers
bs4
firecrawl-py
langchainhub
wikipedia

langchain-community
unstructured
unstructured[all-docs]